third_party/blink/web_tests/external/wpt/css/css-pseudo/input-element-pseudo-open-click.optional-expected.txt

This is a testharness.js-based test.
[FAIL] CSS :open for <input type=date>
  assert_true: Should match :open after opening the picker. expected true got false
[FAIL] CSS :open for <input type=datetime-local>
  assert_true: Should match :open after opening the picker. expected true got false
[FAIL] CSS :open for <input type=week>
  assert_true: Should match :open after opening the picker. expected true got false
[FAIL] CSS :open for <input type=month>
  assert_true: Should match :open after opening the picker. expected true got false
[FAIL] CSS :open for <input type=time>
  assert_true: Should match :open after opening the picker. expected true got false
[FAIL] CSS :open for <input type=text list=datalist>
  assert_true: Should match :open after opening the list. expected true got false
Harness: the test ran to completion.
